Update marketHours.py
Because the call to `api.get_calendar()` returns an array of calendars for the time frame specified, the example needs to index into the array for the print to function properly. As it stands, Python returns an attribute error, as an array containing calendars has no attributes `open` or `close`.